refactor: layers to fillers
