+++ crates/provider/src/fillers/gas.rs
@@ -1,5 +1,5 @@
 use crate::{
-    layers::{FillerControlFlow, TxFiller},
+    fillers::{FillerControlFlow, TxFiller},
     utils::Eip1559Estimation,
     Provider,
 };
