+++ crates/provider/src/lib.rs
@@ -39,7 +39,7 @@ extern crate tracing;
 mod builder;
 pub use builder::{Identity, ProviderBuilder, ProviderLayer, Stack};
 
-pub mod layers;
+pub mod fillers;
 
 mod chain;
